Update braille.scad
Fixed some warnings and deprecations.
Added support for numbers (added the octothorpe (#) for the Braille number sign, underscore (_) for the decimal point and also the digits which are just repeats of a--j.
Added a caret "^" for the Braille capitalization symbol.
Added a flag to allow for left-justified text as well as centered (the default).